soft-sandbox
implementing a main process module which is run for every process. doing things with a wrapper like this means that we have control over what they can do. this probably isn't a very effective sandbox, but the main point is to prevent unintended or easy ways to access the filesystem, which could corrupt it as the OS should be controlling that. The API here sits in the middle to abstract things a bit for processes and make sure theres a consistent way to communicate, e.g. with the filesystem to open a file for reading or writing, or to read stdin etc.